friends donw

+++ database/postgres.js
@@ -201,10 +201,11 @@ module.exports.friendsByID = (id = null) => {
     if (!id) return resolve(null);
     return pool.connect((err, client, done) => {
       if (err) return resolve(null);
-      client.query(`SELECT * FROM users WHERE id IN (SELECT friend_id FROM friends WHERE user_id = ${id})`, (err, result) => {
+      client.query(`SELECT DISTINCT UNNEST(ARRAY[user_id, friend_id]) FROM friends WHERE user_id = ${id} OR friend_id = ${id}`, (err, result) => {
         done();
         if (err) return resolve(null);
-        return resolve(result.rows);
+        const formatted = result.rows.filter(row => row != id);
+        return resolve(formatted);
       });
     });
   });
@@ -215,9 +216,9 @@ module.exports.addFriend = (uid = null, id = null) => {
     if (!id || !uid) return resolve(null);
     return pool.connect((err, client, done) => {
       if (err) return resolve(null);
-      client.query(`SELECT friend_id FROM friends WHERE user_id = ${uid}`, (err, result) => {
+      client.query(`SELECT * FROM friends WHERE (user_id = ${uid} AND friend_id = ${uid}) OR (user_id = ${uid} AND friend_id = ${id})`, (err, result) => {
         if (!err) return resolve(null);
-        client.query(`INSERT INTO friends(user_id, friend_id) VALUES(${uid}, ${id})`, (err, result) => {
+        client.query(`INSERT INTO friends_pending(user_id, friend_id) VALUES(${uid}, ${id})`, (err, result) => {
           if (err) return resolve(null);
           client.query(`SELECT * FROM users WHERE id = ${id}`, (err, result) => {
             done();
@@ -230,14 +231,37 @@ module.exports.addFriend = (uid = null, id = null) => {
   });
 };
 
+module.exports.confirmFriend = (uid = null, id = null) => {
+  return new Promise(resolve => {
+    if (!id || !uid) return resolve(null);
+    return pool.connect((err, client, done) => {
+      if (err) return resolve(null);
+      client.query(`SELECT user_id FROM friends_pending WHERE friend_id = ${uid}`, (err, result) => {
+        if (!err) return resolve(null);
+        client.query(`INSERT INTO friends(user_id, friend_id) VALUES(${id}, ${uid})`, (err, result) => {
+          if (err) return resolve(null);
+          client.query(`DELETE FROM friends_pending WHERE user_id = ${id} AND friend_id = ${uid}`, (err, result) => {
+            if (err) return resolve(null);
+            client.query(`SELECT * FROM users WHERE id = ${id}`, (err, result) => {
+              done();
+              if (err) return resolve(null);
+              return resolve(result.rows);
+            });
+          });
+        });
+      });
+    });
+  });
+};
+
 module.exports.removeFriend = (uid = null, id = null) => {
   return new Promise(resolve => {
     if (!id || !uid) return resolve(null);
     return pool.connect((err, client, done) => {
       if (err) return resolve(null);
-      client.query(`SELECT friend_id FROM friends WHERE user_id = ${uid}`, (err, result) => {
-        if (!err) return resolve(null);
-        client.query(`DELETE FROM friends WHERE user_id = ${uid} AND friend_id = ${id}`, (err, result) => {
+      client.query(`SELECT * FROM friends WHERE (user_id = ${uid} AND friend_id = ${uid}) OR (user_id = ${uid} AND friend_id = ${id})`, (err, result) => {
+        if (err) return resolve(null);
+        client.query(`DELETE FROM friends WHERE (user_id = ${uid} AND friend_id = ${id}) OR (user_id = ${id} AND friend_id = ${uid})`, (err, result) => {
           done();
           if (err) return resolve(null);
           return resolve(result.rows);
